refactor: extract start popup logic
